Pulls out PaddedSwitch
