+++ Blockzilla/Lib/PaddedSwitch.swift
@@ -1,0 +1,18 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import UIKit
+
+class PaddedSwitch: UIView {
+    private static let Padding: CGFloat = 8
+
+    convenience init(switchView: UISwitch) {
+        self.init(frame: .zero)
+
+        addSubview(switchView)
+
+        frame.size = CGSize(width: switchView.frame.width + PaddedSwitch.Padding, height: switchView.frame.height)
+        switchView.frame.origin = CGPoint(x: PaddedSwitch.Padding, y: 0)
+    }
+}
